+++ Source/NakamaNetwork.DB/dbo/Tables/Notifications/Notifications.sql
@@ -2,7 +2,7 @@
 (
     [Id] INT NOT NULL IDENTITY(1,1),
     [UserId] NVARCHAR(450) NOT NULL,
-    [EventType] INT NOT NULL,
+    [EventType] TINYINT NOT NULL,
     [EventId] INT NULL,
     [ReceivedDate] DATETIMEOFFSET(7) NOT NULL CONSTRAINT [DF_dbo.Notifications_ReceivedDate] DEFAULT SYSDATETIMEOFFSET(),
     CONSTRAINT [PK_dbo.Notifications] PRIMARY KEY CLUSTERED ([Id] ASC),
